pyproject.toml
[project]
name = "kosmic-lab"
version = "0.1.0"
description = "Unified research platform for the Kosmic Simulation & Coherence Framework."
authors = [{name = "Kosmic Lab Team", email = "kosmic-lab@example.org"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["coherence", "consciousness", "simulation", "reciprocity", "k-index"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry]
name = "kosmic-lab"
version = "0.1.0"
description = "Unified research platform for the Kosmic Simulation & Coherence Framework."
authors = ["Kosmic Lab Team <kosmic-lab@example.org>"]
packages = [
    { include = "core" },
    { include = "fre" },
    { include = "historical_k" },
    { include = "experiments" }
]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26"
pandas = "^2.1"
networkx = "^3.1"
tqdm = "^4.66"
pyyaml = "^6.0"
jsonschema = "^4.19"
gymnasium = "^0.29"
stable-baselines3 = {version = "^2.2", extras = ["torch"] }
matplotlib = "^3.8"
scikit-learn = "^1.5"
scipy = "^1.12"
seaborn = "^0.13"
requests = "^2.32"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
pytest-xdist = "^3.5"
hypothesis = "^6.92"
pre-commit = "^3.6"
mypy = "^1.7"
black = "^23.12"
isort = "^5.13"
flake8 = "^7.0"
bandit = "^1.7"
sphinx = "^7.2"
sphinx-rtd-theme = "^2.0"
jupyter = "^1.0"
nbconvert = "^7.14"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
